+++ src/encode.cc
@@ -475,7 +475,7 @@ std::vector<uint8_t> EncodeGraph(const UncompressedGraph &g,
     }
   }
 
-  std::vector<float> bits_per_ctx;
+  std::vector<double> bits_per_ctx;
   if (allow_random_access) {
     HuffmanEncode(tokens, kNumContexts, &writer, node_degree_indices,
                   &bits_per_ctx);
@@ -486,27 +486,27 @@ std::vector<uint8_t> EncodeGraph(const UncompressedGraph &g,
   auto stop = std::chrono::high_resolution_clock::now();
 
   if (absl::GetFlag(FLAGS_print_bits_breakdown)) {
-    float degree_bits = 0;
+    double degree_bits = 0;
     for (size_t i = kFirstDegreeContext; i < kReferenceContextBase; i++) {
       degree_bits += bits_per_ctx[i];
     }
-    float reference_bits = 0;
+    double reference_bits = 0;
     for (size_t i = kReferenceContextBase; i < kBlockCountContext; i++) {
       reference_bits += bits_per_ctx[i];
     }
-    float block_bits = 0;
+    double block_bits = 0;
     for (size_t i = kBlockCountContext; i < kFirstResidualBaseContext; i++) {
       block_bits += bits_per_ctx[i];
     }
-    float first_residual_bits = 0;
+    double first_residual_bits = 0;
     for (size_t i = kFirstResidualBaseContext; i < kResidualBaseContext; i++) {
       first_residual_bits += bits_per_ctx[i];
     }
-    float residual_bits = 0;
+    double residual_bits = 0;
     for (size_t i = kResidualBaseContext; i < kNumContexts; i++) {
       residual_bits += bits_per_ctx[i];
     }
-    float total_bits = data.size() * 8.0f;
+    double total_bits = data.size() * 8.0f;
     fprintf(stderr, "Degree bits:         %10.2f [%5.2f bits/edge]\n",
             degree_bits, degree_bits / edges);
     fprintf(stderr, "Reference bits:      %10.2f [%5.2f bits/edge]\n",
